% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{balancesampling}
\alias{balancesampling}
\title{Perform upsampling with SMOTE}
\usage{
balancesampling(
  dat,
  labels,
  topfeaturenumber = 50000,
  cutoff = 10,
  downsampling = FALSE,
  sampleseed = 1234,
  k = 5,
  adjustbetas = TRUE
)
}
\arguments{
\item{dat}{A matrix with features as columns and samples as rows.}

\item{labels}{A vector or factor recording the sample class labels, and the
label order should match the sample order in the \code{dat} matrix.}

\item{topfeaturenumber}{Before performing the upsampling, the most variable
features in the whole data will be selected and others will be dropped, so
that the returned data will only contain these top variable features. The
default value is 50000, meaning the top 50000 variable features will be
selected. If want to kept all the original features, set it as NULL.}

\item{cutoff}{A number to define the small sample classes. Default is 10,
meaning a sample class with < 10 samples will be deemed as a small sample
class and upsampling will be performed on it to synthesize some simulated
samples from its original samples, so that its final sample number can be
equal to 10.}

\item{downsampling}{For the class with a sample number > \code{cutoff}. If
this parameter is set as TRUE, a dnsampling will be performed on them so
that their sample number will be reduced to \code{cutoff}. Default value
is FALSE.}

\item{sampleseed}{Random seed for the sampling processes.}

\item{k}{The number of the nearest neighbors of a sample to synthesize its
SMOTE samples. Default is 5.}

\item{adjustbetas}{If the synthesized samples have beta values <= 0 or >= 1,
this parameter can be set as TRUE, so that these values will be replaced
by the smallest and largest values in the data and also within the range
of (0, 1).}
}
\value{
A list with 2 slots. The slot named "dat" is a matrix containing all
the samples after the upsampling process. The other slot named "labels" is
a vector with the corresponding class labels.
}
\description{
Perform upsampling on small sample classes with SMOTE (Synthetic Minority
Over-sampling Technique).
}
\examples{
library(methylClass)

labels <- system.file('extdata', 'testlabels.rds', package = 'methylClass')
labels <- readRDS(labels)

betas <- system.file('extdata', 'testbetas.rds', package = 'methylClass')
betas <- readRDS(betas)

upsampledbetas <- balancesampling(dat = betas, labels = labels, 
 topfeaturenumber = NULL, cutoff = 100, downsampling = FALSE, 
 sampleseed = 1234, k = 5, adjustbetas = TRUE)
}
