% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{clustergrid}
\alias{clustergrid}
\title{Perform grid search to find the best DBSCAN parameters able to keep samples
with a matching relationship between histological labels and methylation
clusters}
\usage{
clustergrid(
  tsnedat,
  epses = seq(0.25, 5, 0.25),
  minPtses = seq(2, 21, 1),
  cutoff = 0.8,
  classlabels
)
}
\arguments{
\item{tsnedat}{A matrix recording the sample coordinates in a  2-dimensional
tSNE embedding. It should contain 2 columns, and each one corresponds the
sample coordinates of one dimension. The row names are the sample names.}

\item{epses}{The candidate epsilon parameters for DBSCAN clustering on the
tSNE coordinates. It is a numeric vector to be grid searched. Default is
a vector from 0.25 to 5.00, with a stepsize of 0.25.}

\item{minPtses}{The candidate minimum cluster points parameters for DBSCAN
clustering on the tSNE coordinates. It is a numeric vector to be searched.
Default is a vector from 2 to 21, with a stepsize of 1.}

\item{cutoff}{A number between 0 and 1. Default value is 0.8, meaning to
screen for samples with a matching relationship between their histological
labels and DBSCAN clusters, if a DBSCAN cluster has > 80\% of its samples
with the same histological label, it will be kept, while a cluster without
so many dominant samples will be dropped. For the dominant clusters, the
<= 20\% minor samples in them will also be dropped. Thus, the kept clusters
will only contain the dominant samples.}
}
\value{
Different DBSCAN parameters will generate different clusters so that
after the filtering, their kept clusters and samples will be different.
This function will return the parameters with the most samples reserved.
}
\description{
Perform grid search across the parameters epsilon and minimum cluster points
to find the best DBSCAN result to keep the most samples with a matching
relationship between their histological labels and methylation clusters.
}
