% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assistant.R
\name{clustering}
\alias{clustering}
\title{Perform DBSCAN on beta value matrix}
\usage{
clustering(
  y = NULL,
  betas,
  subset.CpGs = 10000,
  seed = 1234,
  cores = 10,
  topfeaturenumber = 50000,
  lasso = 3.25e-07,
  ridge = 0,
  n_pcs = 100,
  perplexity = 10,
  pythonpath = NULL,
  anno = NULL,
  confoundings = NULL,
  padjcut = 0.05,
  xcutoff = 0.1,
  cutnum = 10000,
  eps = 3.5,
  minPts = 5,
  legendcolnum = 2
)
}
\arguments{
\item{y}{The true labels of the samples. Can be a vector, factor, or NULL.
If it is a vector or factor, each element is a label for a sample and the
element order in it should be the same as the sample order in the sample
data provided by the parameter \code{betas}. This is necessary only if the
DBSCAN clustering is based on features selected via the \code{limma}, and
in other cases, it can be set as NULL. Default is NULL.}

\item{betas}{The beta value matrix of the samples. Each row is one sample
and each column is one feature.}

\item{subset.CpGs}{The feature selection method. It can be a numeric number
such as 10000, and then the top 10000 most variable features of the data
will be selected. It can also be the string "limma", so that \code{limma}
will be used to select the significantly differential features between a
sample class and all other samples. The differential ones should fulfill
the condition that their adjusted p-value < \code{padjcut} (default value
is 0.05), and the betas value difference between the sample class and all
other samples should be > \code{xcutoff}, or < -\code{xcutoff} (default is
0.1). After the differential features for each class have been selected by
\code{limma}, they will be mixed together and ordered according to their
adjusted p-value and the absolute of beta value difference, and then the
top \code{cutnum} ones (default number is 10000) will be selected and as
the final features. Additionally, if there are any confounding factors in
the dataset need to be removed, they should be provided via the parameter
\code{confoundings} and \code{limma} will select the features after these
confoundings have been adjusted. The parameter \code{confoundings} should
be provided with a vector with the confounding names in the meta data (the
meta data are provided via the parameter \code{anno}) as elements. As to
\code{subset.CpGs}, it can also be the string "SCMER", and then the method
\code{SCMER} will be used to select features from the data. In this case,
other parameters should be set well, including \code{lasso}, \code{ridge},
\code{n_pcs}, \code{perplexity}, \code{savefigures}, \code{pythonpath} and
\code{topfeaturenumber}. Because \code{SCMER} is to select features able
to preserve the original manifold structure of the data after its feature
selection by elastic net, most of these parameters are used to config the
manifold (\code{n_pcs} and \code{perplexity}) and elastic net processes
(\code{lasso} and \code{ridge}), while another important parameter is the
\code{pythonpath}, which is used to tell the function the absolute path of
the \code{Python} interpreter you want to use to run \code{SCMER} because
this method also depends on \code{Python}. Besides, because \code{limma}
and \code{SCMER} can be time-consuming if running on large data, it is
recommended to do a prescreen on the data before running them, and the
parameter \code{topfeaturenumber} can be set as a numeric value such as
50000, so that the top 50000 most variable features will be selected, and
the top variable, \code{limma}, or the \code{SCMER} features can then be
selected further on the prescreened data.}

\item{seed}{The tSNE step before DBSCAN clustering involve a random process
and a random seed is needed to be set here to ensure the result can be
repeated. Default value is 1234.}

\item{cores}{The core number need to do parallelization computation. Default
is 10.}

\item{topfeaturenumber}{As mentioned in the \code{subset.CpGs} parameter
part, it is used to set the prescreened feature number. Default is 50000.
It can also be set as NULL, so that no precreen will be done on the data.}

\item{lasso}{A parameter special for \code{SCMER} feature selection and it
defines the strength of L1 regularization in the elastic net process of
\code{SCMER}. Default is 3.25e10-7, so that around 10000 features will be
selected from 50000 prescreened candidate features.}

\item{ridge}{A parameter special for \code{SCMER} feature selection and it
defines the strength of L2 regularization in the elastic net process of
\code{SCMER}. Default is 0, so that the elastic net process is actually
a LASSO process.}

\item{n_pcs}{Number of principle components need to reconstruct the sample-
sample distance matrix during the \code{SCMER} selection. In addition, it
also defines the number of priciple components need to do tSNE before the
DBSCAN clustering. Default value is 100.}

\item{perplexity}{Perplexity of tSNE modeling. Default is 10.}

\item{pythonpath}{Because the feature selection method \code{SCMER} is a
\code{Python} based method, the directory of the \code{Python} interpreter
you want to use to run it should be transferred to the function via this
parameter, and several \code{Python} modules need to be installed to your
\code{Python} environment, including \code{time}, \code{functiontools},
\code{abc}, \code{torch}, \code{typing}, \code{sklearn}, \code{pandas},
\code{numpy}, \code{matplotlib}, \code{multiprocessing}, \code{scanpy}.}

\item{anno}{A data frame recording the meta data of the samples, and should
contain at least 2 columns named as "label" and "sentrix". The former one
records the sample labels while the latter one records the sample IDs that
also used as row names of the methylation data matrix. The default value
is NULL and it is not necessary, but if need to use \code{limma} to do the
feature selection and remove the confoundings, it should be provided with
the confounding factors included in it.}

\item{confoundings}{A parameter special for \code{limma} feature selection.
Details can be seen in the \code{subset.CpGs} parameter section.}

\item{padjcut}{A parameter for \code{limma} feature selection. Default value
is 0.05 and details can be seen in the \code{subset.CpGs} section.}

\item{xcutoff}{A parameter for \code{limma}. Details can also be seen in the
\code{subset.CpGs} section. Default value is 0.1.}

\item{cutnum}{A parameter special for \code{limma}. Details can be seen in
the \code{subset.CpGs} section.}

\item{eps}{Size (radius) of the epsilon neighborhood set for DBSCAN, default
is 3.5}

\item{minPts}{Number of minimum points required in the eps neighborhood for
core points (including the point itself) in DBSCAN. Default is 5.}

\item{legendcolumn}{The column number of the figure legend for the scatter
plot generated. Default is 2.}
}
\value{
Will return a beta value matrix containing the samples attributed to
different DBSCAN clusters and with the ones out of the clusters removed,
and also a data frame annotating the cluster IDs of these samples will be
returned with the beta value matrix.
}
\description{
Perform density-based spatial clustering of applications with noise (DBSCAN)
on beta value matrix
}
