% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning.R
\name{makecv}
\alias{makecv}
\title{Divide samples into different cross validation loops}
\usage{
makecv(y, cv.fold = 5, seednum = 1234, normalcv = FALSE)
}
\arguments{
\item{y}{The true labels of the samples. Can be a vector or factor. Each of
its element is a label for a sample and the element order in it should be
the same as the sample order in the sample data provided to the function
\code{maincv} to do cross validation for classification models.}

\item{cv.fold}{A numeric number and the default value is 5, so that if the
parameter \code{normalcv} is TRUE, a 5 fold cross validation structure
will be created, and if \code{normalcv} is FALSE, a 5 by 5 nested cross
validation structure will be created.}

\item{seednum}{The random seed for the sampling process to assign samples to
different cross validation loops. Default is 1234.}

\item{normalcv}{Indicating whether the cross validation loops are normal or
nested. Default is FALSE, meaning nested cross validation.}
}
\value{
Will return a multiple layer list recording the samples assigned to
each cross validation loop and the samples are indicated by their indeces
in \code{y}.
}
\description{
Create a cross validation structure and assign samples to different loops
}
