% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylClass.R
\docType{package}
\name{methylClass}
\alias{methylClass}
\title{methylClass: an R package to construct DNA methylation-based classification
models}
\description{
The R package methylClass is special for performing DNA methylation-based
classification. Within it, we provide the eSVM (ensemble-based support
vector machine) model to achieve much higher accuracy in methylation data
classification than the popular random forest (RF) model and overcome the
time-consuming problem of the traditional SVM. In addition, some novel
feature selection methods, such as SCMER, are included in the package to
improve the classification. Furthermore, we also provide functions for
multi-omics studies.
}
\section{methylClass functions}{

balancesampling
clustergrid
clustering
cvdata
labelclusters
mainJvisR
maincalibration
maincv
mainfeature
mainpredict
maintrain
makecv
multipledistribution
toDMR
togene
}

